specify absolute path of compiler envs
ensures that enabling toolchain specifies a specific compiler on the system, and not the one that may be installed in the env with the same name, unless explicitly requested.

This is the right thing to do for halide (https://github.com/conda-forge/staged-recipes/pull/2559#discussion_r105329430), but may not be what people want in general. When people install gcc as a build dep (and toolchain), do they expect the one in the env to be used by default without being explicitly selected? Should they?