Use a compiler wrapper
This uses the compiler wrappers implemented in spack package manager
and creates a new toolchain package. Link here. Macports has a similar
idea in their tracker, https://trac.macports.org/ticket/44413

There is a bash script cc installed in $PREFIX/bin/conda_forge/ that
looks at the arguments passed to it and process them before sending
it to the real cc compiler. There are symlinks to gcc, clang, g++,
gfortran etc.

This script looks at what mode the compiler/linker is invoked in and
sets the mode. 'mode' is set to one of:

    vcheck  version check
    cpp     preprocess
    cc      compile
    as      assemble
    ld      link
    ccld    compile & link

CONDA_FORGE_CFLAGS is sent to the compiler if cc and ccld.
CONDA_FORGE_CPPFLAGS which includes -I${PREFIX}/include is sent in
 cc, ccld, as, cpp.
CONDA_FORGE_LDFLAGS which includes -L${PREFIX}/lib is sent in ld, ccld.
rpath is sent in ccld and ld modes with -Wl,-rpath and -rpath.
This avoids issues we've seen in toolchain package.

  numpy.distutils packages like scipy.
     Setting LDFLAGS env variable overwrites the LDFLAGS variable in
     numpy.distutils instead of appending. In toolchain-v3 we are not
     setting LDFLAGS at all, but sending them directly to the compiler
     and linker.

  rpy2
     rpy2's build system checks CFLAGS and preprocess them and chokes
     when -headerpad_max_install_names is set. toolchain-v3 does not
     have this problem.

This package also supports setting CFLAGS and CXXFLAGS which are sent
to the compiler through the build system.