Fix file tests on Windows
Given the change in install location of the Windows activate/deactivate
scripts from `%LIBRARY_PREFIX%` to `%PREFIX%`, the file tests also need
to be updated to make sure they landed in the right place. This fixes
the Windows files tests to check for the activate/deactivate scripts in
`%PREFIX%` where they should be.