Set `SDKROOT` for CMake on macOS
Make sure the `SDKROOT` environment variable is set for CMake on macOS
to point to the correct location macOS SDK. This effectively the same as
adding the `-isysroot` flag to the compiler flags. However if CMake
strips the existing `-isysroot` flag that we add from the compiler flags
for some reason, this ensures CMake still finds the macOS SDK we wish to
use and adds the appropriate `-isysroot` flag as well. CMake will pick
this up from the `SDKROOT` environment variable if it is not otherwise
specified in the flags. More details in the docs link below.

Also we pass along `CMAKE_OSX_SYSROOT` as an environment variable as
well so that users can easily specify the flag if they would like to be
a bit more careful. Though it shouldn't be needed.

ref: https://cmake.org/cmake/help/v3.12/variable/CMAKE_OSX_SYSROOT.html